requirements/base.txt
asgiref==3.2.7
Django==3.0.6
django-filter==2.2.0
django-model-utils==4.0.0
djangorestframework==3.11.0
drf-flex-fields==0.8.5
drf-writable-nested==0.6.0
psycopg2-binary==2.8.5
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
